Utilize HTTPS instead of HTTP.
HTTP is a insecure protocol as everything is sent in the open with no transport security.

This also bumps Chartist.js from 0.9.4 to 0.11.4.

There are a few dead links in these JavaScript plugins. 

We need to in a future pull request go through all of our included plugins/libraries and remove anything that is not being utilized.